+++ post_office/__init__.py
@@ -1,4 +1,4 @@
-VERSION = (2, 0, 8)
+VERSION = (3, 0, 4)
 
 from .backends import EmailBackend